Fix OTA build failure due to refactoring
