+++ libraries/freertos_plus/aws/ota/src/aws_iot_ota_pal.h
@@ -26,7 +26,7 @@
 #ifndef _AWS_OTA_PAL_H_
 #define _AWS_OTA_PAL_H_
 
-#include "aws_ota_types.h"
+#include "aws_iot_ota_types.h"
 #include "aws_iot_ota_agent.h"
 
 /**
