+++ libraries/freertos_plus/aws/ota/test/aws_ota_agent_test_access_declare.h
@@ -33,9 +33,9 @@
 #ifndef _AWS_OTA_AGENT_TEST_ACCESS_DECLARE_H_
 #define _AWS_OTA_AGENT_TEST_ACCESS_DECLARE_H_
 
-#include "aws_ota_types.h"
+#include "aws_iot_ota_types.h"
 #include "aws_iot_ota_agent.h"
-#include "aws_ota_agent_internal.h"
+#include "aws_iot_ota_agent_internal.h"
 
 IngestResult_t TEST_OTA_prvIngestDataBlock( OTA_FileContext_t * C,
                                             const char * pacRawMsg,
