+++ libraries/freertos_plus/aws/ota/test/aws_test_ota_pal.c
@@ -35,7 +35,7 @@
 #include "unity_fixture.h"
 #include "unity.h"
 #include "aws_ota_pal_test_access_declare.h"
-#include "aws_ota_pal.h"
+#include "aws_iot_ota_pal.h"
 #include "aws_iot_ota_agent.h"
 #include "iot_pkcs11.h"
 #include "iot_pkcs11_config.h"
