+++ vendors/microchip/boards/curiosity_pic32mzef/ports/ota/aws_ota_pal.c
@@ -28,7 +28,7 @@
 
 /*lint -e9045 Ignore advisories about non-hidden definitions in header files. */
 
-#include "aws_ota_pal.h"
+#include "aws_iot_ota_pal.h"
 #include "iot_pkcs11.h"
 #include "aws_nvm.h"
 #include "iot_crypto.h"
