+++ vendors/pc/boards/windows/aws_demos/config_files/aws_ota_agent_config.h
@@ -86,14 +86,14 @@
   */
 #define otaconfigMAX_NUM_BLOCKS_REQUEST        1U
 
-  /**
-   * @brief The maximum number of requests allowed to send without a response before we abort.
-   *
-   * This configuration parameter sets the maximum number of times the requests are made over
-   * the selected communication channel before aborting and returning error.
-   *
-   */
+/**
+ * @brief The maximum number of requests allowed to send without a response before we abort.
+ *
+ * This configuration parameter sets the maximum number of times the requests are made over
+ * the selected communication channel before aborting and returning error.
+ *
+ */
 
-#define otaconfigMAX_NUM_REQUEST_MOMENTUM       32U        
+#define otaconfigMAX_NUM_REQUEST_MOMENTUM       32U
 
 #endif /* _AWS_OTA_AGENT_CONFIG_H_ */
