+++ vendors/ti/boards/cc3220_launchpad/aws_tests/config_files/aws_ota_agent_config.h
@@ -73,16 +73,25 @@
 
 /**
  * @brief The maximum number of data blocks requested from OTA streaming service.
- * 
- *  This configuration parameter is sent with data requests and represents the maximum number of 
- *  data blocks the service will send in response. The maximum limit for this must be calculated 
- *  from the maximum data response limit (128 KB from service) divided by the block size. 
+ *
+ *  This configuration parameter is sent with data requests and represents the maximum number of
+ *  data blocks the service will send in response. The maximum limit for this must be calculated
+ *  from the maximum data response limit (128 KB from service) divided by the block size.
  *  For example if block size is set as 1 KB then the maximum number of data blocks that we can
- *  request is 128/1 = 128 blocks. Configure this parameter to this maximum limit or lower based on 
- *  how many data blocks response is expected for each data requests. 
+ *  request is 128/1 = 128 blocks. Configure this parameter to this maximum limit or lower based on
+ *  how many data blocks response is expected for each data requests.
  *  Please note that this must be set larger than zero.
- *  
+ *
  */
  #define otaconfigMAX_NUM_BLOCKS_REQUEST        8U
+
+/**
+ * @brief The maximum number of requests allowed to send without a response before we abort.
+ *
+ * This configuration parameter sets the maximum number of times the requests are made over
+ * the selected communication channel before aborting and returning error.
+ *
+ */
+#define otaconfigMAX_NUM_REQUEST_MOMENTUM       32U
 
 #endif /* _AWS_OTA_AGENT_CONFIG_H_ */
